[v0.89.1][maintainability] Split cli/run_artifacts/cognitive.rs by responsibility (Closes #1944)
